+++ bootstrap.php
@@ -21,7 +21,7 @@
 // Require 3rdparty autoloader
 require('3rdparty/autoload.php');
 
-/**
+/*
  * PSR-4 autoloader
  *
  * @param string $class the fully-qualified class name