+++ src/Cache/CacheAware.php
@@ -20,6 +20,8 @@
 
 namespace YoutubeDownloader\Cache;
 
+use Psr\SimpleCache\CacheInterface;
+
 /**
  * Describes a cache-aware instance
  */
@@ -27,6 +29,10 @@ interface CacheAware
 {
     /**
      * Sets a cache instance on the object
+     *
+     * @deprecated since version 0.8, setCache() must be a Psr\SimpleCache\CacheInterface instance as argument 1 in 0.9
+     *
+     * param Psr\SimpleCache\CacheInterface $cache
      *
      * @param Cache $cache
      */