+++ src/Cache/CacheException.php
@@ -21,12 +21,13 @@
 namespace YoutubeDownloader\Cache;
 
 use Exception;
+use Psr\SimpleCache\CacheException as ExceptionInterface;
 
 /**
  * Base exception in a cache.
  *
  * This interface must be compatible with PSR-16 Psr\SimpleCache\CacheException
  */
-class CacheException extends Exception
+class CacheException extends Exception implements ExceptionInterface
 {
 }