+++ src/Cache/NullCache.php
@@ -21,12 +21,13 @@
 namespace YoutubeDownloader\Cache;
 
 use DateTimeInterface;
+use Psr\SimpleCache\CacheInterface;
 
 /**
  * A cache instance that does nothing
  */
 
-class NullCache implements Cache
+class NullCache implements Cache, CacheInterface
 {
     /**
      * Fetches a value from the cache.
@@ -110,5 +111,87 @@ class NullCache implements Cache
                 '`a-z`, `A-Z`, `0-9`, `_` and `-`'
             ));
         }
+    }
+
+    /**
+     * Wipes clean the entire cache's keys.
+     *
+     * @return bool true on success and false on failure
+     */
+    public function clear()
+    {
+        throw new \Exception(__METHOD__ . ' is not implemented');
+    }
+
+    /**
+     * Obtains multiple cache items by their unique keys.
+     *
+     * @param iterable $keys    a list of keys that can obtained in a single operation
+     * @param mixed    $default default value to return for keys that do not exist
+     *
+     * @throws \Psr\SimpleCache\InvalidArgumentException
+     *                                                   MUST be thrown if $keys is neither an array nor a Traversable,
+     *                                                   or if any of the $keys are not a legal value
+     *
+     * @return iterable A list of key => value pairs. Cache keys that do not exist or are stale will have $default as value.
+     */
+    public function getMultiple($keys, $default = null)
+    {
+        throw new \Exception(__METHOD__ . ' is not implemented');
+    }
+
+    /**
+     * Persists a set of key => value pairs in the cache, with an optional TTL.
+     *
+     * @param iterable               $values a list of key => value pairs for a multiple-set operation
+     * @param null|int|\DateInterval $ttl    Optional. The TTL value of this item. If no value is sent and
+     *                                       the driver supports TTL then the library may set a default value
+     *                                       for it or let the driver take care of that.
+     *
+     * @throws \Psr\SimpleCache\InvalidArgumentException
+     *                                                   MUST be thrown if $values is neither an array nor a Traversable,
+     *                                                   or if any of the $values are not a legal value
+     *
+     * @return bool true on success and false on failure
+     */
+    public function setMultiple($values, $ttl = null)
+    {
+        throw new \Exception(__METHOD__ . ' is not implemented');
+    }
+
+    /**
+     * Deletes multiple cache items in a single operation.
+     *
+     * @param iterable $keys a list of string-based keys to be deleted
+     *
+     * @throws \Psr\SimpleCache\InvalidArgumentException
+     *                                                   MUST be thrown if $keys is neither an array nor a Traversable,
+     *                                                   or if any of the $keys are not a legal value
+     *
+     * @return bool True if the items were successfully removed. False if there was an error.
+     */
+    public function deleteMultiple($keys)
+    {
+        throw new \Exception(__METHOD__ . ' is not implemented');
+    }
+
+    /**
+     * Determines whether an item is present in the cache.
+     *
+     * NOTE: It is recommended that has() is only to be used for cache warming type purposes
+     * and not to be used within your live applications operations for get/set, as this method
+     * is subject to a race condition where your has() will return true and immediately after,
+     * another script can remove it making the state of your app out of date.
+     *
+     * @param string $key the cache item key
+     *
+     * @throws \Psr\SimpleCache\InvalidArgumentException
+     *                                                   MUST be thrown if the $key string is not a legal value
+     *
+     * @return bool
+     */
+    public function has($key)
+    {
+        throw new \Exception(__METHOD__ . ' is not implemented');
     }
 }